requirements.txt
einops>=0.4.1
kornia>=0.6.5
pytorch-lightning>=1.7.7
matplotlib>=3.5.2
opencv-python>=4.6.0.66
imageio~=2.29.0
imageio-ffmpeg
scipy~=1.10.1
tensorboard
torch~=2.0.1+cu118
numpy~=1.24.1
tqdm~=4.65.0
taichi-nightly